change filename

+++ analysis/secondlevel.py
@@ -52,7 +52,7 @@ gmThresh=args.gmThresh
 
 
 mask_img=threshold_img('/home/jmuraskin/Projects/CCD/working_v1/seg_probabilities/grey_matter_mask.nii.gz',threshold=gmThresh)
-mask_name='/home/jmuraskin/Projects/CCD/working_v1/seg_probabilities/grey_matter_mask-%d-percent.nii.gz' % gmThresh*100
+mask_name='/home/jmuraskin/Projects/CCD/working_v1/seg_probabilities/grey_matter_mask-%d-percent.nii.gz' % int(gmThresh*100)
 mask_img.to_filename(mask_name)
 
 if runFC:
